requirements.txt
boto3
google-cloud-storage
azure-storage-blob